# chunk:imporve append bytes in chunk