@coocood 2018-01-22
@zhexuany 
This is an optimization for a specific case, and it can be much worse for other cases.