@zhexuany 2018-02-03
I came up with two approaches:
1. take control over for extending array 
2. use bytes.Buffer 

After some benchmarks, both has some performance penalty. Hence, closing PR.